requirements.txt
flask
flask-socketio
requests
bs4
selenium
yt-dlp
moviepy<2.0
pillow
transformers
openai
whisper
certifi
gunicorn
gevent
imageio_ffmpeg
imageio
pandas